feat: implemen esm modules task
