+++ src/modules/esm.mjs
@@ -1,0 +1,51 @@
+import { readFile } from 'node:fs/promises';
+import path from 'node:path';
+import url from 'node:url';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'node:http';
+import './files/c.js';
+
+const __filename = url.fileURLToPath(new URL(import.meta.url));
+const __dirname = path.dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject;
+if (random > 0.5) {
+    unknownObject = JSON.parse(await readFile(new URL('./files/a.json', import.meta.url)));
+} else {
+    unknownObject = JSON.parse(await readFile(new URL('./files/b.json', import.meta.url)));
+}
+/** variant 2:
+if (random > 0.5) {
+    unknownObject = await import('./files/a.json', { assert: { type: 'json' } });
+} else {
+    unknownObject = await import('./files/b.json', { assert: { type: 'json' } });
+}
+unknownObject = unknownObject.default;
+ */
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
